@@ -9,23 +9,24 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@tabler/icons-react": "^3.5.0",
-    "firebase": "^10.12.2",
+    "@tabler/icons-react": "^3.31.0",
+    "autoprefixer": "^10.4.16",
+    "firebase": "^10.14.1",
     "next": "14.2.3",
     "next-themes": "^0.3.0",
     "prettier-plugin-organize-imports": "^3.2.4",
     "prettier-plugin-tailwindcss": "^0.5.14",
-    "react": "^18",
-    "react-dom": "^18"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.1",
+    "postcss": "^8.4.31"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
+    "@types/node": "^20.17.24",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "eslint": "^8.57.1",
     "eslint-config-next": "14.2.3",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5.8.2"
   }
 }
