@@ -1,9 +1,13 @@
-const defaultTheme = require("tailwindcss/defaultTheme");
+import type { Config } from "tailwindcss";
+import defaultTheme from "tailwindcss/defaultTheme";
 
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./src/**/*.{js,ts,jsx,tsx,mdx}"],
-  darkMode: ["selector", '[data-mode="dark"]'],
+const config: Config = {
+  content: [
+    "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+  ],
+  darkMode: ["class", '[data-mode="dark"]'],
   theme: {
     colors: {
       transparent: "transparent",
@@ -24,3 +28,5 @@ module.exports = {
   },
   plugins: [],
 };
+
+export default config;
